+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniApp.kt
@@ -169,10 +169,15 @@ abstract class MiniApp internal constructor() {
      * Get the manifest information e.g. required and optional permissions.
      * @param appId mini app id.
      * @param versionId of mini app.
-     * @return MiniAppManifest an object contains manifest information of a miniapp.
+     * @param languageCode of mini app.
+     * @return MiniAppManifest an object contains manifest information of a mini app.
      */
     @Throws(MiniAppSdkException::class)
-    abstract suspend fun getMiniAppManifest(appId: String, versionId: String): MiniAppManifest
+    abstract suspend fun getMiniAppManifest(
+        appId: String,
+        versionId: String,
+        languageCode: String = ""
+    ): MiniAppManifest
 
     /**
      * Get the currently downloaded manifest information e.g. required and optional permissions.