+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/TestData.kt
@@ -27,6 +27,7 @@ internal const val TEST_MA_DISPLAY_NAME = "test_name"
 internal const val TEST_MA_ICON = "test_icon"
 internal const val TEST_MA_VERSION_TAG = "test_vtag"
 internal const val TEST_MA_VERSION_ID = "test_vid"
+internal const val TEST_MA_PREVIEW_CODE = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
 
 internal const val TEST_HA_NAME = "test_hostapp_name"
 internal const val TEST_HA_ID_PROJECT = "test_project_id"
@@ -76,3 +77,5 @@ internal val TEST_HA_ANALYTICS_CONFIGS = listOf(TEST_CONFIG1, TEST_CONFIG2)
 
 internal const val TEST_PUBLIC_KEY_ID = "test_public_key_id"
 internal const val TEST_MANIFEST_SIGNATURE = "test_manifest_signature"
+
+internal const val TEST_LANG_MANIFEST_DEFAULT = "default"