+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.9.0 (2021-XX-XX)
 **SDK**
+- **Feature:** Added `languageCode` parameter in `MiniApp.getMiniAppManifest` to support for internationalized manifest.
 - **Feature:** Added `hostLocale` in `HostEnvironmentInfo` to provide default language value from Host App.
 
 **Sample App**